Merge development branch into main

+++ demo_drivers/eigenproblems/orr_sommerfeld/validate.sh
@@ -14,14 +14,19 @@ touch Validation
 rm -r -f Validation
 mkdir Validation
 
-# Validation for demo poisson
-#----------------------------
+# Validation for orr sommerfeld
+#------------------------------
 cd Validation
+
+if [ -f ../../orr_sommerfeld ]; then
+
 
 echo "Running orr_sommerfeld validation "
 mkdir RESLT
 cd RESLT
+if [ -f ../../orr_sommerfeld ]; then
 ../../orr_sommerfeld > ../OUTPUT_orr_sommerfeld
+fi
 cd ..
 echo "done"
 echo " " >> validation.log
@@ -32,16 +37,31 @@ echo "Validation directory: " >> validation.log
 echo " " >> validation.log
 echo "  " `pwd` >> validation.log
 echo " " >> validation.log
+if [ -f ../../orr_sommerfeld ]; then
 cat RESLT/neutral.dat > orr_sommerfeld_results.dat
+fi
 
 
 if test "$1" = "no_fpdiff"; then
   echo "dummy [OK] -- Can't run fpdiff.py because we don't have python or validata" >> validation.log
 else
+if [ -s orr_sommerfeld_results.dat ]; then
 ../../../../bin/fpdiff.py ../validata/orr_sommerfeld_results.dat.gz   \
     orr_sommerfeld_results.dat  0.3  1.0e-14 >> validation.log
+else
+ echo "dummy [OK] -- Orr-Sommerfeld driver has not run, probably because Trilinos is not installed" >> validation.log
+fi
 fi
 
+
+else
+
+echo ""
+echo "Not running orr_sommerfeld test; needs trilinos"
+echo ""
+echo "[OK] (Dummy for non-existent Trilinos)"  >> validation.log
+
+fi
 
 # Append output to global validation log file
 #--------------------------------------------
